Adding copies of Game
These copies are named for lab 11 and lab 12, they are identical ( except for the lab name ) to game.py.  I wanted to leave game.py around however for historical ref if needed.  Otherwise these will be what we turn in.  Lab 11, will need to be slimmed down a bit since it has everything for lab 11 as well as 12 that is a superset.